`into_iter` against a `Vector` yields values
Previous commits invoked `into_iter` on an array.

This [Stack Overflow post](https://stackoverflow.com/questions/34733811/what-is-the-difference-between-iter-and-into-iter)
states "Arrays implement `IntoIterator` (in such a surprising fashion)
to make it possible to iterate over references to them in `for` loops."

This commit changes the type of `haystack` from `Array` to `Vector`.
Then everything works "as advertised."